instructions ADC, AND, ASL and NOP are implemented and fully supported ; I'm gonna keep implementing instructions ; should make tests w bigger RAM at some point ; when the VM will be working I'll make some more advanced test using online test benches but I need to implement the instructions first

+++ src/decode.h
@@ -6,8 +6,8 @@
 #include "memory.h"
 
 
-#define CLEAR_FLAG(cpu, flag) {cpu->p &= ~(flag);}
-#define RAISE_FLAG(cpu, flag) {cpu->p |= (flag);}
+#define CLEAR_FLAG(cpu, flag) {cpu->p &= (byte)~(flag);}
+#define RAISE_FLAG(cpu, flag) {cpu->p |= (byte)(flag);}
 
 #define COND_RAISE_FLAG(cpu, flag, cond) {if(cond){ RAISE_FLAG(cpu, flag);}else {CLEAR_FLAG(cpu, flag);}}
